requirements.txt
google-genai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.64.0
scikit-learn>=1.3.0
pyyaml>=6.0
rich>=13.0.0
pathlib2>=2.3.7
pytest>=7.4.0
pre-commit>=3.4.0 